Rename to globalmodel
