+++ src/main/java/org/visab/globalmodel/cbrshooter/PlayerStatistics.java
@@ -1,4 +1,4 @@
-package org.visab.generalmodelchangeme.cbrshooter;
+package org.visab.globalmodel.cbrshooter;
 
 public class PlayerStatistics {
