+++ src/main/java/org/visab/globalmodel/cbrshooter/ShooterDataRepresentation.java
@@ -1,4 +1,4 @@
-package org.visab.generalmodelchangeme.cbrshooter;
+package org.visab.globalmodel.cbrshooter;
 
 /**
  * Renamed from "StatisticsForPathViewer to ShooterDataRepresentation
