+++ src/main/java/org/visab/globalmodel/cbrshooter/Vector2.java
@@ -1,4 +1,4 @@
-package org.visab.generalmodelchangeme.cbrshooter;
+package org.visab.globalmodel.cbrshooter;
 
 public class Vector2 {
