+++ src/main/java/org/visab/processing/cbrshooter/CBRShooterListener.java
@@ -5,9 +5,9 @@ import java.util.UUID;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.visab.generalmodelchangeme.cbrshooter.CBRShooterFile;
-import org.visab.generalmodelchangeme.cbrshooter.CBRShooterMapImage;
-import org.visab.generalmodelchangeme.cbrshooter.CBRShooterStatistics;
+import org.visab.globalmodel.cbrshooter.CBRShooterFile;
+import org.visab.globalmodel.cbrshooter.CBRShooterMapImage;
+import org.visab.globalmodel.cbrshooter.CBRShooterStatistics;
 import org.visab.processing.ReplaySessionListenerBase;
 import org.visab.util.AssignByGame;
