+++ src/main/java/org/visab/processing/cbrshooter/ShooterParser.java
@@ -6,7 +6,7 @@ import java.io.IOException;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.visab.generalmodelchangeme.cbrshooter.ShooterDataRepresentation;
+import org.visab.globalmodel.cbrshooter.ShooterDataRepresentation;
 import org.visab.processing.AbstractParser;
 
 /**
