+++ src/main/java/org/visab/repository/DatabaseRepository.java
@@ -1,7 +1,7 @@
 package org.visab.repository;
 
-import org.visab.generalmodelchangeme.IVISABFile;
-import org.visab.generalmodelchangeme.VISABFileBase;
+import org.visab.globalmodel.IVISABFile;
+import org.visab.globalmodel.VISABFileBase;
 import org.visab.util.AssignByGame;
 import org.visab.util.JsonConvert;
 import org.visab.util.Settings;
