+++ src/main/java/org/visab/util/AssignByGame.java
@@ -6,13 +6,13 @@ import java.util.UUID;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.visab.generalmodelchangeme.IStatistics;
-import org.visab.generalmodelchangeme.IVISABFile;
-import org.visab.generalmodelchangeme.cbrshooter.CBRShooterFile;
-import org.visab.generalmodelchangeme.cbrshooter.CBRShooterMapImage;
-import org.visab.generalmodelchangeme.cbrshooter.CBRShooterStatistics;
-import org.visab.generalmodelchangeme.starter.DefaultFile;
-import org.visab.generalmodelchangeme.starter.DefaultStatistics;
+import org.visab.globalmodel.IStatistics;
+import org.visab.globalmodel.IVISABFile;
+import org.visab.globalmodel.cbrshooter.CBRShooterFile;
+import org.visab.globalmodel.cbrshooter.CBRShooterMapImage;
+import org.visab.globalmodel.cbrshooter.CBRShooterStatistics;
+import org.visab.globalmodel.starter.DefaultFile;
+import org.visab.globalmodel.starter.DefaultStatistics;
 import org.visab.processing.IMapImage;
 import org.visab.processing.ISessionListener;
 import org.visab.processing.cbrshooter.CBRShooterListener;
